requirements.txt
requests==2.31.0
yfinance==0.2.44
alpaca-trade-api==3.2.0
python-dotenv==1.0.1
pandas==2.2.0
flask==3.1.0
